feat: 预设二级菜单+5峰PID自整定+编码器ISR采样+AHT20非阻塞10Hz等
- UI: 耗材预设改为二级菜单(编辑/新增/删除/退出)，列表>5项出滚动条，长按切换当前预设；预设名称唯一身份仅新建可编辑；编辑页单退出+改动才确认保存；确认弹窗中文标题居中；时间/温度弹窗样式对齐
- PID: 元件/空气两套自整定改5峰ZN法(进度每峰20%，99%写库→100%驻留)，校准页实时显示AIR/PTC温度；PID调整页按入口区分元件组/空气组
- 编码器: 相位改1kHz SysTick中断采样不丢步+旋转加速
- AHT20: 非阻塞状态机10Hz采样、上电40ms延时、初始化校准位校验+软复位重试
- CS1237: 4帧中值滤波、DOUT上拉、量程拒收、自动去皮仅开机窗口(防烘干中误清物料)
- 安全: 温升监控锚定初始温度+1℃裕度防抖，防启动误报箱体破损
- RGB: 开机随背光渐亮启动灯效
- 外部Flash引入SFUD驱动；README补充特性/预设/PID自整定说明

+++ module/mod_web_server.c
@@ -152,8 +152,8 @@ static void SendSensorData(int conn_id)
     uint16_t hours = (uint16_t)(secs / 3600);
     uint16_t mins = (uint16_t)((secs % 3600) / 60);
     uint16_t secs_rem = (uint16_t)(secs % 60);
-    sprintf(json, "{\"temp\":%.1f,\"hum\":%.1f,\"time\":\"%02d:%02d:%02d\",\"weight\":%.1f}",
-            g_sys.current_temp, g_sys.current_humidity, hours, mins, secs_rem, g_sys.weight_g);
+    sprintf(json, "{\"temp\":%.1f,\"hum\":%.1f,\"time\":\"%02d:%02d:%02d\",\"weight\":%d}",
+            g_sys.current_temp, g_sys.current_humidity, hours, mins, secs_rem, (int)g_sys.weight_g);
     (void)json;
 }
